requirements.txt
# local package
-e .

# external requirements
#click
#Sphinx
#coverage
#awscli
#flake8
#python-dotenv>=0.5.1

# Additional requirements
antlr4-python3-runtime==4.9.3
certifi==2022.6.15
charset-normalizer==2.1.1
cycler==0.11.0
fonttools==4.37.1
hydra-core==1.2.0
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
omegaconf==2.2.3
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
torch==1.12.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
